refactor: nested graph.
