+++ common-ui/src/main/java/com/benidict/common_ui/navigation/graph/CartNavGraph.kt
@@ -7,12 +7,12 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.benidict.common_ui.navigation.route.CartGraph
-import com.benidict.common_ui.navigation.route.CartScreeRoute
+import com.benidict.common_ui.navigation.route.CartScreenRoute
 
 fun NavGraphBuilder.cartNavGraph(navRoute: @Composable (NavBackStackEntry, route: Any) -> Unit) {
-    navigation<CartGraph>(startDestination = CartScreeRoute) {
-        composable<CartScreeRoute> {
-            navRoute(it, CartScreeRoute)
+    navigation<CartGraph>(startDestination = CartScreenRoute) {
+        composable<CartScreenRoute> {
+            navRoute(it, CartScreenRoute)
         }
     }
 }
